tutorex/requirements.txt
asgiref==3.3.4
Django~=3.2
django-phonenumber-field==5.1.0
django-utils-six==2.0
phonenumbers==8.12.21
pytz==2021.1
sqlparse==0.4.1
django-heroku
gunicorn
psycopg2
psycopg2-binary
whitenoise
pip~=21.1
dj-database-url